Improve support for splitted packages when DEVELSRCDIR is set.
An example is 'linux-git', which is the Linux kernel source on AUR.
This is a splitted package, which contains a large git clone. It would
be nice if it doesn't have to be downloaded completely each time,
which is the essence of the DEVELSRCDIR (DEVELBUILDDIR) option.

This patch assumes that the correct pkgname name is in the first
element of the array. If pkgname is not an array, the normal
pkgname will be used, which was the behaviour before this patch.