Change default installation path when running make install
Currently when running 'make install' yaourt will install itself
into /usr/local. Maybe I don't understand why, but if I want to
install yaourt it should go to /usr and the configuration file to
/etc.

This commit changes the default installation path for these to. If
this is unwished behaviour, please drop this commit.