Add --ignore function to the man documentation
When I wanted to update my software with yaourt (yaourt -Syua),
everything got updated as the tool should do. This is great,
however one of my packages is 'linux-git' and on my computer this
takes about 4-6 hours of compilation time.

So while being on my way to implement an --except or --exclude
option for the sync, I noticed that pacman already had this
function implemented as --ignore. After a couple of tests, I
noticed that this functionality also worked in yaourt.

The reason why I did not use this option is because it was not
documented in the manual pages. Therefore, this commit, adding
this simple but awesome feature.

The rest of all the commits are just a bonus. :)